refactor: extract input component from message log
Also remove chatId prop
